refactor: introduce ConstraintSolverInterface
